# Conflicts:
#	lib/src/main/kotlin/viewPart/nodes/drawableBIN/BINDrawableTree.kt
#	lib/src/main/kotlin/viewPart/nodes/drawableRB/RBDrawableNode.kt
#	lib/src/main/kotlin/viewPart/nodes/drawableTree/DrawableNode.kt
#	lib/src/main/kotlin/viewPart/nodes/drawableTree/DrawableTree.kt

+++ lib/src/main/kotlin/databaseManage/AVLTreeManager.kt
@@ -37,7 +37,7 @@ class AVLTreeManager : TreeManager<
         for (el in drawVertex) ans.add(
             DrawableAVLVertex(
                 value = el.value,
-                height = el.height.toInt(),
+                height = el.height,
                 x = -0.0,
                 y = -0.0
             )
